Add new code of Dashboard.jsx and Logout.jsx files to frontend/src/views/auth
Use rfce (from ES7+  extension downloaded and installed) to create Dashboard.jsx and Logout.jsx pages.  In Logout, provide links to register and login pages. In Dashboard, if user is logged in, provide link to the logout page otherwise provide links to register and login pages. Finally, In App.jsx (in frontend/src) append routes or paths to dashboard and logout pages. Test if all links are working as expected and if access and refresh tokens are being generated (in Cookies) after logging in. Also test if said tokens are getting cleared (not visible in Cookies) after logging out (try refreshing the page to be sure)