Configure the features - size, color and qty of a product detail page
At the beginning of src/views/store/ProductDetail.jsx in the function, create variables for size, color and quantity while fetching their values from the database to be shown here. Initially, show quantity as 1, size as "No Size" and Color as "No Color" and change them as user interacts with the items. "Add to cart" button for now only logs values of quantity, size and color to console for now.