Additions to Products.jsx to dynamically show changes to quantity, size and color
Make Additions to the code in Products.jsx to dynamically show changes to quantity, size and color - on screen and in console (to test). Initial values of color and size are "No Color" and "No Size" respectively. As we make changes, print them to console to test and finally display the changes on screen.