Fetch cart from API and change UI based on number of items in cart
Get cart id and user data from imported functions and call cart-list url (from backend/backend/api/urls.py) and write fetchCartData function to visit the url and store the data in a variable. If cart id is not null or undefined, check if userData is available and call the fetchCartData function accordingly. Populate the UI of /cart/ url based on the obtained variables.