Add class CartDetailView in backend/store/views.py and add new url in backend/api/urls.py
In CartDetailView, we already know about serializer_class and permission_classes but now, lookup_field is used. lookup_field is set to card_id. By default lookup_field is primary key of a view/table in the database but here we set it to cart_id. We are familiar with the get_queryset function but here we write the get function which calls the get_queryset to get the cart items. We now iterate over the cart items to sum up each key price value (so total of shipping, tax, service_fee and sub_total is stored in total) and we return response i.e. to frontend when this class' get function is called. Finally, we add a new url cart-detail in backend/api/urls.py file.