Update Cart.jsx to update values without reloading as we change number of items
Add toast and the function updateCart() in Cart.jsx to update Sub Total and Cart Summary as we change the number of items in cart and press the refresh button (in blue). Mention onClick in the blue refresh button to call updateCart method. Finally, test the changes in the UI itself.