Add CartItemDeleteAPIView and add urls to call the view
Add CartItemDeleteAPIView in backend/store/views.py with serializer CartSerializer and lookup_field cart_id. Again, if user exists, we get that user's cart id and item id and return (to delete) and if user doesn't, we get just the item id and cart id and return. Add urls (for existing user and anon user) in backend/api/urls.py which are directed to the recently created class.