Write new function in Cart.jsx to delete items as and when Remove is clicked
Write handleDeleteCartItem function in Cart.jsx to form the url to delete cart item based on whether user is logged in or not. We had written code in the backend views.py (CartItemDeleteAPIView) just returning cart, now we delete item from the frontend in Cart.jsx using apiInstance.delete(url) after forming the url. Call the handleDeleteCartItem function when Remove button is clicked.